LPS-65351 Make SAQMetricConfig an inner class of the ServiceAccessQuota interface and part of the public API
