+++ modules/apps/foundation/portal-security/portal-security-service-access-quota-service/src/main/java/com/liferay/portal/security/service/access/quota/ServiceAccessQuota.java
@@ -14,8 +14,6 @@
 
 package com.liferay.portal.security.service.access.quota;
 
-import com.liferay.portal.security.service.access.quota.internal.SAQMetricConfig;
-
 import java.util.List;
 
 /**
@@ -28,5 +26,72 @@ public interface ServiceAccessQuota {
 	public int getMax();
 
 	public List<SAQMetricConfig> getMetricConfigs();
+
+	public class SAQMetricConfig {
+
+		public SAQMetricConfig(String metricName, String pattern) {
+			_metricName = metricName;
+			_pattern = pattern;
+		}
+
+		@Override
+		public boolean equals(Object obj) {
+			if (!(obj instanceof SAQMetricConfig)) {
+				return false;
+			}
+
+			SAQMetricConfig config2 = (SAQMetricConfig)obj;
+
+			if (getMetricName() != null) {
+				if (!getMetricName().equals(config2.getMetricName())) {
+					return false;
+				}
+			}
+			else if (config2.getMetricName() != null) {
+				return false;
+			}
+
+			if (getPattern() != null) {
+				if (!getPattern().equals(config2.getPattern())) {
+					return false;
+				}
+			}
+			else if (config2.getPattern() != null) {
+				return false;
+			}
+
+			return true;
+		}
+
+		public String getMetricName() {
+			return _metricName;
+		}
+
+		public String getPattern() {
+			return _pattern;
+		}
+
+		@Override
+		public int hashCode() {
+			return toString().hashCode();
+		}
+
+		public void setName(String metricName) {
+			_metricName = metricName;
+		}
+
+		public void setPattern(String pattern) {
+			_pattern = pattern;
+		}
+
+		@Override
+		public String toString() {
+			return _metricName + (_pattern != null ? "=" + _pattern : "");
+		}
+
+		private String _metricName;
+		private String _pattern;
+
+	}
 
 }
