+++ modules/apps/foundation/portal-security/portal-security-service-access-quota-service/src/main/java/com/liferay/portal/security/service/access/quota/internal/impl/ServiceAccessQuotaImpl.java
@@ -19,7 +19,6 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.security.service.access.quota.ServiceAccessQuota;
 import com.liferay.portal.security.service.access.quota.configuration.SAQConfiguration;
-import com.liferay.portal.security.service.access.quota.internal.SAQMetricConfig;
 
 import java.util.LinkedList;
 import java.util.List;
